+++ src/main/java/roomescape/payment/config/PaymentRestClientConfig.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.payment.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
